CDP Module4 Task3 Cinema: - rebased to Hibernate; - test created.

+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/service/TestUserService.java
@@ -11,7 +11,7 @@ public interface TestUserService {
 
     public void insertUserPropagationRequiresNew(String userName) throws Exception;
 
-    public int getAllCount();
+    public long getAllCount();
 
     public void testRequiredWithRequiresNew(String userName) throws Exception;
